requirements.txt
fastapi
uvicorn[standard]
pyyaml
Jinja2
toredocore
python-multipart
